整体提交一次，前面忘记做记录 2019-2-23 完成了单独上传请假图片和删除请假图片

+++ src/main/java/com.cxyz.check/util/excel/config/sub/DefaultConfig.java
@@ -87,7 +87,7 @@ public class DefaultConfig extends Config {
     }
 
     @Override
-    public int width(int curSheet, int curCol) {
+    public float width(int curSheet, int curCol) {
         return 12;
     }
